requirements.txt
asttokens==2.0.7
attrs==22.1.0
backcall==0.2.0
cattrs==23.1.2
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
colorlog==6.6.0
configobj==5.0.6
contourpy==1.1.0
cplot==0.9.3
cryptography==41.0.1
cycler==0.11.0
Cython==0.29.32
dbus-python==1.2.16
decorator==5.1.1
exceptiongroup==1.1.1
executing==0.10.0
fjson==0.1.6
fonttools==4.40.0
h5py==3.9.0
h5sparse==0.1.0
idna==2.8
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==1.1.1
ipython==8.4.0
jedi==0.17.2
joblib==1.3.1
kiwisolver==1.4.4
llvmlite==0.40.1
markdown-it-py==3.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.3
matplotx==0.3.10
mdurl==0.1.2
meshio==5.3.5
mpmath==1.3.0
ndim==0.1.21
networkx==3.1
npx==0.1.1
ntplib==0.4.0
numba==0.57.1
numpy==1.23.1
orthopy==0.9.7
packaging==21.3
pandas==2.0.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
Pint==0.21.1
platformdirs==3.8.0
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-machineid==0.3.0
pyaml==23.12.0
pycparser==2.21
Pygments==2.15.1
PyGObject==3.36.0
pymoab==5.1.1
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.2
pytest-runner==6.0.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-license-manager==0.2.6
pytz==2023.3
PyYAML==5.3.1
requests==2.22.0
requests-cache==1.1.0
requests-unixsocket==0.2.0
rich==13.4.2
rich-argparse==1.1.1
scikit-learn==1.3.0
scipy==1.9.0
shapely==2.0.3
six==1.14.0
stack-data==0.4.0
sympy==1.12
threadpoolctl==3.1.0
tomli==2.0.1
tomli-w==1.0.0
traitlets==5.3.0
typing-extensions==4.7.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==1.25.8
wcwidth==0.2.5
winregistry==1.1.1
x21==0.4.12
zipp==3.15.0
